+++ api-reports/2_13.txt
@@ -13985,6 +13985,42 @@ NodeSelector[JT] def querySelector(selectors: String): Element
 NodeSelector[JT] def querySelectorAll(selectors: String): NodeList[Node]
 NonDocumentTypeChildNode[JT] def nextElementSibling: Element
 NonDocumentTypeChildNode[JT] def previousElementSibling: Element
+Notification[JC] def addEventListener[T <: Event](`type`: String, listener: js.Function1[T, _], options: EventListenerOptions): Unit
+Notification[JC] def addEventListener[T <: Event](`type`: String, listener: js.Function1[T, _], useCapture: Boolean?): Unit
+Notification[JC] val body: String
+Notification[JC] def close(): Unit
+Notification[JC] val data: js.Object
+Notification[JC] val dir: String
+Notification[JC] def dispatchEvent(evt: Event): Boolean
+Notification[JC] val icon: String
+Notification[JC] val lang: String
+Notification[JC] val noscreen: Boolean
+Notification[JC] var onclick: js.Function0[Any]
+Notification[JC] var onerror: js.Function0[Any]
+Notification[JC] def removeEventListener[T <: Event](`type`: String, listener: js.Function1[T, _], options: EventListenerOptions): Unit
+Notification[JC] def removeEventListener[T <: Event](`type`: String, listener: js.Function1[T, _], useCapture: Boolean?): Unit
+Notification[JC] val renotify: Boolean
+Notification[JC] val silent: Boolean
+Notification[JC] val sound: String
+Notification[JC] val sticky: Boolean
+Notification[JC] val tag: String
+Notification[JC] val vibrate: js.Array[Double]
+Notification[JO] val permission: String
+Notification[JO] def requestPermission(callback: js.Function1[String, Any]): Unit
+NotificationOptions[JT] val body: String
+NotificationOptions[JT] val dir: String
+NotificationOptions[JT] val icon: String
+NotificationOptions[JT] val lang: String
+NotificationOptions[JT] val noscreen: Boolean
+NotificationOptions[JT] val onclick: js.Function0[Any]
+NotificationOptions[JT] val onerror: js.Function0[Any]
+NotificationOptions[JT] val renotify: Boolean
+NotificationOptions[JT] val silent: Boolean
+NotificationOptions[JT] val sound: String
+NotificationOptions[JT] val sticky: Boolean
+NotificationOptions[JT] val tag: String
+NotificationOptions[JT] val vibrate: js.Array[Double]
+NotificationOptions[SO] def apply(body: js.UndefOr[String] = js.undefined, dir: js.UndefOr[String] = js.undefined, icon: js.UndefOr[String] = js.undefined, lang: js.UndefOr[String] = js.undefined, noscreen: js.UndefOr[Boolean] = js.undefined, renotify: js.UndefOr[Boolean] = js.undefined, silent: js.UndefOr[Boolean] = js.undefined, sound: js.UndefOr[String] = js.undefined, sticky: js.UndefOr[Boolean] = js.undefined, tag: js.UndefOr[String] = js.undefined, onclick: js.UndefOr[js.Function0[Any]] = js.undefined, onerror: js.UndefOr[js.Function0[Any]] = js.undefined, vibrate: js.UndefOr[js.Array[Double]] = js.undefined): NotificationOptions
 OfflineAudioCompletionEvent[JT] def bubbles: Boolean
 OfflineAudioCompletionEvent[JT] def cancelBubble: Boolean
 OfflineAudioCompletionEvent[JT] def cancelable: Boolean
@@ -23237,42 +23273,6 @@ experimental/HttpMethod[SO] val PATCH = "PATCH".asInstanceOf[HttpMethod]
 experimental/HttpMethod[SO] val POST = "POST".asInstanceOf[HttpMethod]
 experimental/HttpMethod[SO] val PUT = "PUT".asInstanceOf[HttpMethod]
 experimental/HttpMethod[SO] val QUERY = "QUERY".asInstanceOf[HttpMethod]
-experimental/Notification[JC] def addEventListener[T <: Event](`type`: String, listener: js.Function1[T, _], options: EventListenerOptions): Unit
-experimental/Notification[JC] def addEventListener[T <: Event](`type`: String, listener: js.Function1[T, _], useCapture: Boolean?): Unit
-experimental/Notification[JC] val body: String
-experimental/Notification[JC] def close(): Unit
-experimental/Notification[JC] val data: js.Object
-experimental/Notification[JC] val dir: String
-experimental/Notification[JC] def dispatchEvent(evt: Event): Boolean
-experimental/Notification[JC] val icon: String
-experimental/Notification[JC] val lang: String
-experimental/Notification[JC] val noscreen: Boolean
-experimental/Notification[JC] var onclick: js.Function0[Any]
-experimental/Notification[JC] var onerror: js.Function0[Any]
-experimental/Notification[JC] def removeEventListener[T <: Event](`type`: String, listener: js.Function1[T, _], options: EventListenerOptions): Unit
-experimental/Notification[JC] def removeEventListener[T <: Event](`type`: String, listener: js.Function1[T, _], useCapture: Boolean?): Unit
-experimental/Notification[JC] val renotify: Boolean
-experimental/Notification[JC] val silent: Boolean
-experimental/Notification[JC] val sound: String
-experimental/Notification[JC] val sticky: Boolean
-experimental/Notification[JC] val tag: String
-experimental/Notification[JC] val vibrate: js.Array[Double]
-experimental/Notification[JO] val permission: String
-experimental/Notification[JO] def requestPermission(callback: js.Function1[String, Any]): Unit
-experimental/NotificationOptions[JT] val body: String
-experimental/NotificationOptions[JT] val dir: String
-experimental/NotificationOptions[JT] val icon: String
-experimental/NotificationOptions[JT] val lang: String
-experimental/NotificationOptions[JT] val noscreen: Boolean
-experimental/NotificationOptions[JT] val onclick: js.Function0[Any]
-experimental/NotificationOptions[JT] val onerror: js.Function0[Any]
-experimental/NotificationOptions[JT] val renotify: Boolean
-experimental/NotificationOptions[JT] val silent: Boolean
-experimental/NotificationOptions[JT] val sound: String
-experimental/NotificationOptions[JT] val sticky: Boolean
-experimental/NotificationOptions[JT] val tag: String
-experimental/NotificationOptions[JT] val vibrate: js.Array[Double]
-experimental/NotificationOptions[SO] def apply(body: js.UndefOr[String] = js.undefined, dir: js.UndefOr[String] = js.undefined, icon: js.UndefOr[String] = js.undefined, lang: js.UndefOr[String] = js.undefined, noscreen: js.UndefOr[Boolean] = js.undefined, renotify: js.UndefOr[Boolean] = js.undefined, silent: js.UndefOr[Boolean] = js.undefined, sound: js.UndefOr[String] = js.undefined, sticky: js.UndefOr[Boolean] = js.undefined, tag: js.UndefOr[String] = js.undefined, onclick: js.UndefOr[js.Function0[Any]] = js.undefined, onerror: js.UndefOr[js.Function0[Any]] = js.undefined, vibrate: js.UndefOr[js.Array[Double]] = js.undefined): NotificationOptions
 experimental/PointerLock[SO] def exitPointerLock(): Unit
 experimental/PointerLock[SO] def movementX: Double
 experimental/PointerLock[SO] def movementY: Double
@@ -23871,9 +23871,12 @@ experimental/mediastream/package[SO] type TextTrackMode = String
 experimental/package[SO] type BodyInit = Blob | BufferSource | FormData | String
 experimental/package[SO] type ByteString = String
 experimental/package[SO] type HeadersInit = Headers | Sequence[Sequence[ByteString]] | OpenEndedDictionary[ByteString]
+experimental/package[SO] type Notification = dom.Notification  (@deprecated in 2.0.0)
+experimental/package[SO] type NotificationOptions = dom.NotificationOptions  (@deprecated in 2.0.0)
 experimental/package[SO] type OpenEndedDictionary[T] = js.Dictionary[T]
 experimental/package[SO] type RequestInfo = String | Request
 experimental/package[SO] type Sequence[T] = js.Array[T]
+experimental/package[SO] lazy val Notification: dom.Notification.type  (@deprecated in 2.0.0)
 experimental/permissions/package[SO] def apply(permissionName: PermissionName): PermissionDescriptor
 experimental/permissions/package[SO] def apply(permissionUserVisibleOnly: Boolean): PushPermissionDescriptor
 experimental/permissions/package[SO] val denied = "denied".asInstanceOf[PermissionState]
