+++ src/main/scala/org/scalajs/dom/experimental/serviceworkers/ServiceWorkers.scala
@@ -1,11 +1,10 @@
 package org.scalajs.dom.experimental.serviceworkers
 
-import org.scalajs.dom.experimental.{
-  Notification, NotificationOptions, Request, Response, Sequence
-}
+import org.scalajs.dom.experimental.{Request, Response, Sequence}
 import org.scalajs.dom.webgl.RenderingContext
 import org.scalajs.dom.{
-  Event, EventInit, EventTarget, MessageEvent, MessagePort, WorkerGlobalScope
+  Event, EventInit, EventTarget, MessageEvent, MessagePort, Notification,
+  NotificationOptions, WorkerGlobalScope
 }
 import scala.scalajs.js
 import scala.scalajs.js.annotation._
